Days of Love/requirements.txt
streamlit
Pillow